requirements.txt
Django==3.1
djangorestframework==3.14.0
psycopg2>=2.8.6,<2.9
drf-spectacular==0.26.2